requirements.txt
streamlit
opencv-python
numpy
pandas
Pillow
scikit-image
scikit-learn
torch
torchvision
onnxruntime
tqdm
sqlalchemy
python-dotenv
google-generativeai
openpyxl
plotly
groq
fpdf2